pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pr_commenter"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "BSD 3-Clause License"}
classifiers = [
    "Programming Language :: Python :: 3",
]
authors = [
    {name = "Martín Gaitán", email = "marting@shiphero.com"}
]
dependencies = [
    "pygithub >= 1.55",
    "docopt-ng >= 0.8.1",
]
dynamic = ["version", "description"]

[project.scripts]
pr_commenter = "pr_commenter:main"


[project.optional-dependencies]
dev = [
    "pytest",
    "black"
]

[tool.ipdb]
context=5


[tool.black]
line-length = 120
target-version = ["py37"]
color = true
